+++ src/collection/collectionToJson.ts
@@ -7,7 +7,7 @@ import type { PipelineCollection } from './PipelineCollection';
  * Note: Functions `collectionToJson` and `createCollectionFromJson` are complementary
  */
 export async function collectionToJson(collection: PipelineCollection): Promise<Array<PipelineJson>> {
-    const promptbookUrls = await collection.listPipelines();
-    const promptbooks = await Promise.all(promptbookUrls.map((url) => collection.getPipelineByUrl(url)));
+    const pipelineUrls = await collection.listPipelines();
+    const promptbooks = await Promise.all(pipelineUrls.map((url) => collection.getPipelineByUrl(url)));
     return promptbooks;
 }