+++ src/conversion/utils/parseCommand.ts
@@ -1,16 +1,18 @@
 import { spaceTrim } from 'spacetrim';
 import { SyntaxError } from '../../errors/SyntaxError';
-import type { Command } from '../../types/Command';
-import type { ExecuteCommand } from '../../types/Command';
-import type { ExpectAmountCommand } from '../../types/Command';
-import type { ExpectCommand } from '../../types/Command';
-import type { ExpectFormatCommand } from '../../types/Command';
-import type { JokerCommand } from '../../types/Command';
-import type { ModelCommand } from '../../types/Command';
-import type { ParameterCommand } from '../../types/Command';
-import type { PostprocessCommand } from '../../types/Command';
-import type { PromptbookUrlCommand } from '../../types/Command';
-import type { PromptbookVersionCommand } from '../../types/Command';
+import type {
+    Command,
+    ExecuteCommand,
+    ExpectAmountCommand,
+    ExpectCommand,
+    ExpectFormatCommand,
+    JokerCommand,
+    ModelCommand,
+    ParameterCommand,
+    PostprocessCommand,
+    PromptbookUrlCommand,
+    PromptbookVersionCommand,
+} from '../../types/Command';
 import { ExecutionTypes } from '../../types/ExecutionTypes';
 import { EXPECTATION_UNITS } from '../../types/PipelineJson/PromptTemplateJson';
 import type { string_markdown_text } from '../../types/typeAliases';
@@ -71,10 +73,10 @@ export function parseCommand(listItem: string_markdown_text): Command {
             );
         }
 
-        const promptbookUrlString = listItemParts.pop()!;
-        const promptbookUrl = new URL(promptbookUrlString);
+        const pipelineUrlString = listItemParts.pop()!;
+        const pipelineUrl = new URL(pipelineUrlString);
 
-        if (promptbookUrl.protocol !== 'https:') {
+        if (pipelineUrl.protocol !== 'https:') {
             throw new SyntaxError(
                 spaceTrim(
                     `
@@ -88,7 +90,7 @@ export function parseCommand(listItem: string_markdown_text): Command {
             );
         }
 
-        if (promptbookUrl.hash !== '') {
+        if (pipelineUrl.hash !== '') {
             throw new SyntaxError(
                 spaceTrim(
                     `
@@ -105,7 +107,7 @@ export function parseCommand(listItem: string_markdown_text): Command {
 
         return {
             type: 'PROMPTBOOK_URL',
-            promptbookUrl,
+            pipelineUrl,
         } satisfies PromptbookUrlCommand;
     } else if (type.startsWith('PROMPTBOOK_VERSION') || type.startsWith('PTBK_VERSION')) {
         if (listItemParts.length !== 2) {