+++ src/types/execution-report/ExecutionReportJson.ts
@@ -1,8 +1,6 @@
 import type { PromptResult } from '../../execution/PromptResult';
 import type { Prompt } from '../Prompt';
-import type { string_markdown_text } from '../typeAliases';
-import type { string_pipeline_url } from '../typeAliases';
-import type { string_version } from '../typeAliases';
+import type { string_markdown_text, string_pipeline_url, string_version } from '../typeAliases';
 
 /**
  * ExecutionReport is result of executing one promptbook
@@ -16,9 +14,9 @@ import type { string_version } from '../typeAliases';
  */
 export type ExecutionReportJson = {
     /**
-     * Unique identifier of the promptbook from promptbook which was executed
+     * Unique identifier of the pipeline from promptbook which was executed
      */
-    readonly promptbookUrl?: string_pipeline_url;
+    readonly pipelineUrl?: string_pipeline_url;
 
     /**
      * Title of from promptbook which was executed
@@ -47,7 +45,7 @@ export type ExecutionReportJson = {
         /**
          * The prompt wich was executed
          */
-        prompt: Omit<Prompt, 'promptbookUrl' | 'parameters'>;
+        prompt: Omit<Prompt, 'pipelineUrl' | 'parameters'>;
 
         /**
          * Result of the prompt execution (if not failed during LLM execution)